Fold DockerContainerBackend into ContainerService
DockerContainerBackend was a thin shim over DockerCLI whose only real
responsibilities were app-name to container-name mapping and entitlement
to docker-flag translation. Both of those are app-lifecycle concerns
that already belong on ContainerService, so routing them through a
dedicated actor added indirection without isolating anything meaningful.

ContainerService now talks to DockerCLI directly and owns those two
helpers itself, which removes a layer and keeps the Linux-container
path readable in one place.
